# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  fullName: string;
  phone: string;
  blockNumber: string;
  buildingNumber: string;
  apartmentNumber: string;
  scheduledAt: string;
  status: BookingStatus;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBookingBody {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 4 */
  phone: string;
  /** @minLength 1 */
  blockNumber: string;
  /** @minLength 1 */
  buildingNumber: string;
  /** @minLength 1 */
  apartmentNumber: string;
  scheduledAt: string;
  notes?: string | null;
}

export interface UpdateBookingStatusBody {
  status: BookingStatus;
}

export interface QueuePosition {
  bookingId: number;
  status: BookingStatus;
  /** Position in queue (1 = next). 0 if not in queue. */
  position: number;
  totalWaiting: number;
  /** Booking ID currently being served, if any */
  currentlyServing?: number | null;
}

export interface QueueSummary {
  totalWaiting: number;
  inProgress: number;
  completedToday: number;
}

export interface AdminStats {
  totalBookings: number;
  pendingCount: number;
  inProgressCount: number;
  completedCount: number;
  cancelledCount: number;
  todayBookings: number;
}

export type GallerySection =
  (typeof GallerySection)[keyof typeof GallerySection];

export const GallerySection = {
  styles: "styles",
  customers: "customers",
} as const;

export interface GalleryItem {
  id: number;
  section: GallerySection;
  imageUrl: string;
  title?: string | null;
  styleType?: string | null;
  createdAt: string;
}

export interface CreateGalleryItemBody {
  section: GallerySection;
  /** @minLength 1 */
  imageUrl: string;
  title?: string | null;
  styleType?: string | null;
}

export type ListBookingsParams = {
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListGalleryItemsParams = {
  section?: ListGalleryItemsSection;
};

export type ListGalleryItemsSection =
  (typeof ListGalleryItemsSection)[keyof typeof ListGalleryItemsSection];

export const ListGalleryItemsSection = {
  styles: "styles",
  customers: "customers",
} as const;