# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new booking
 */
export const createBookingBodyFullNameMin = 2;

export const createBookingBodyPhoneMin = 4;

export const CreateBookingBody = zod.object({
  fullName: zod.string().min(createBookingBodyFullNameMin),
  phone: zod.string().min(createBookingBodyPhoneMin),
  blockNumber: zod.string().min(1),
  buildingNumber: zod.string().min(1),
  apartmentNumber: zod.string().min(1),
  scheduledAt: zod.coerce.date(),
  notes: zod.string().nullish(),
});

/**
 * @summary List all bookings (admin)
 */
export const ListBookingsQueryParams = zod.object({
  status: zod
    .enum(["pending", "in_progress", "completed", "cancelled"])
    .optional(),
});

export const ListBookingsResponseItem = zod.object({
  id: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  blockNumber: zod.string(),
  buildingNumber: zod.string(),
  apartmentNumber: zod.string(),
  scheduledAt: zod.coerce.date(),
  status: zod.enum(["pending", "in_progress", "completed", "cancelled"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListBookingsResponse = zod.array(ListBookingsResponseItem);

/**
 * @summary Get a booking
 */
export const GetBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookingResponse = zod.object({
  id: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  blockNumber: zod.string(),
  buildingNumber: zod.string(),
  apartmentNumber: zod.string(),
  scheduledAt: zod.coerce.date(),
  status: zod.enum(["pending", "in_progress", "completed", "cancelled"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update booking status (admin)
 */
export const UpdateBookingStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBookingStatusBody = zod.object({
  status: zod.enum(["pending", "in_progress", "completed", "cancelled"]),
});

export const UpdateBookingStatusResponse = zod.object({
  id: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  blockNumber: zod.string(),
  buildingNumber: zod.string(),
  apartmentNumber: zod.string(),
  scheduledAt: zod.coerce.date(),
  status: zod.enum(["pending", "in_progress", "completed", "cancelled"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Get the queue position and status for a booking
 */
export const GetQueuePositionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetQueuePositionResponse = zod.object({
  bookingId: zod.number(),
  status: zod.enum(["pending", "in_progress", "completed", "cancelled"]),
  position: zod
    .number()
    .describe("Position in queue (1 = next). 0 if not in queue."),
  totalWaiting: zod.number(),
  currentlyServing: zod
    .number()
    .nullish()
    .describe("Booking ID currently being served, if any"),
});

/**
 * @summary Overall queue summary
 */
export const GetQueueSummaryResponse = zod.object({
  totalWaiting: zod.number(),
  inProgress: zod.number(),
  completedToday: zod.number(),
});

/**
 * @summary Admin dashboard stats
 */
export const GetAdminStatsResponse = zod.object({
  totalBookings: zod.number(),
  pendingCount: zod.number(),
  inProgressCount: zod.number(),
  completedCount: zod.number(),
  cancelledCount: zod.number(),
  todayBookings: zod.number(),
});

/**
 * @summary List gallery items
 */
export const ListGalleryItemsQueryParams = zod.object({
  section: zod.enum(["styles", "customers"]).optional(),
});

export const ListGalleryItemsResponseItem = zod.object({
  id: zod.number(),
  section: zod.enum(["styles", "customers"]),
  imageUrl: zod.string(),
  title: zod.string().nullish(),
  styleType: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListGalleryItemsResponse = zod.array(ListGalleryItemsResponseItem);

/**
 * @summary Add a gallery item (admin)
 */

export const CreateGalleryItemBody = zod.object({
  section: zod.enum(["styles", "customers"]),
  imageUrl: zod.string().min(1),
  title: zod.string().nullish(),
  styleType: zod.string().nullish(),
});

/**
 * @summary Delete a gallery item (admin)
 */
export const DeleteGalleryItemParams = zod.object({
  id: zod.coerce.number(),
});