# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 4 */
  phone: string;
  /** @minLength 1 */
  blockNumber: string;
  /** @minLength 1 */
  buildingNumber: string;
  /** @minLength 1 */
  apartmentNumber: string;
  scheduledAt: Date;
  notes?: string | null;
}