# lib/api-zod/src/generated/types/createGalleryItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */
import type { GallerySection } from "./gallerySection";

export interface CreateGalleryItemBody {
  section: GallerySection;
  /** @minLength 1 */
  imageUrl: string;
  title?: string | null;
  styleType?: string | null;
}