# lib/api-zod/src/generated/types/galleryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */
import type { GallerySection } from "./gallerySection";

export interface GalleryItem {
  id: number;
  section: GallerySection;
  imageUrl: string;
  title?: string | null;
  styleType?: string | null;
  createdAt: Date;
}