# lib/api-zod/src/generated/types/gallerySection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */

export type GallerySection =
  (typeof GallerySection)[keyof typeof GallerySection];

export const GallerySection = {
  styles: "styles",
  customers: "customers",
} as const;