# lib/api-zod/src/generated/types/queuePosition.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface QueuePosition {
  bookingId: number;
  status: BookingStatus;
  /** Position in queue (1 = next). 0 if not in queue. */
  position: number;
  totalWaiting: number;
  /** Booking ID currently being served, if any */
  currentlyServing?: number | null;
}