# lib/api-zod/src/generated/types/queueSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */

export interface QueueSummary {
  totalWaiting: number;
  inProgress: number;
  completedToday: number;
}