# lib/api-zod/src/generated/types/updateBookingStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qassa booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface UpdateBookingStatusBody {
  status: BookingStatus;
}